Update that Makes it Work
Fixed syntax in the HTML Bootstrap script tags so that Bootstrap would load.
Modified colors to be the exact colors of UW and UIUC.
Properly made navbar with the .navbar class and made the rest of the page sink to accomodate.
Made footer with contact information
Added picture of me in the top right corner.
Made some small adjustments to font sizing and color in the table, navbar, and footer.