Character Summary Working
Some Linting Changes for myself

Added a resultant array of <p> elements that give an overview of the class and level that have been selected.

Currently, once the "Submit Level" button is pushed, the array is generated, and cannot be changed unless you reload the page.

I need to add in an override or add-on method to do the same for spellcaster, but with spells per day. I also need to override or add-on for the class features that change with level.

Should I make all of these renderable React Components? Probably.